==> data/step1_clean.ipynb <==
## Preliminaries

import pandas as pd
import numpy as np
from geopy.geocoders import Nominatim
geolocator = Nominatim()

## Clean MyWorld Survey Data

xls_file = pd.ExcelFile('raw_data/latestmyworld-women_2016_07_05_15_44_10.xlsx')

xls_df = xls_file.parse(xls_file.sheet_names[0])

df = pd.DataFrame()

#q1
df['Belief In Earning Opportunity Equality'] = xls_df['q1']
#q2
df['Belief In Pay Equality'] = xls_df['q2']

#q3/1
df['Education Helps Women'] = xls_df['q3/1']
#q3/2
df['Role Models Help Women'] = xls_df['q3/2']
#q3/3
df['Family Planning Helps Women'] = xls_df['q3/3']
#q3/4
df['Maternity Pay And Childcare Helps Women'] = xls_df['q3/4']
#q3/5
df['Minimum Wage Help Women'] = xls_df['q3/5']
#q3/6
df['Access To Banking Helps Women'] = xls_df['q3/6']
#q3/7
df['Womens Groups And Charities Helps Women'] = xls_df['q3/7']
#q3/a
df['What Helps Women?'] = xls_df['q3a']

#q4/1
df['Fear Of Violence Is A Barrier'] = xls_df['q4/1']
#q4/2
df['Laws Against Women Working Are A Barrier'] = xls_df['q4/2']
#q4/3
df['Family Responsibilities Are A Barrier'] = xls_df['q4/3']
#q4/4
df['Lack Of Experience Or Skills Is A Barrier'] = xls_df['q4/4']
#q4/5
df['Traditional Views Of Women Are A Barrier'] = xls_df['q4/5']
#q4/6
df['Lack Of ICT Access Is A Barrier'] = xls_df['q4/6']
#q4/a
df['What Are Barriers To Women?'] = xls_df['q4a']

#q5
df['Strength Of Belief That Government Is Helping Women'] = xls_df['q5']
#q6
df['What Helped You The Most?'] = xls_df['q6']

#gender
df['Gender'] = xls_df['respondent/gender']

#education
df['Education'] = xls_df['respondent/edu']

#age
df['Age'] = xls_df['respondent/age']

#country
df['Country'] = xls_df['respondent/country']

#city
df['City'] = xls_df['respondent/qd5']

#location
df['Location'] = np.nan

# Geolocate (city, country). If the city doesn't exist, just geolocate the country.

geo = []

for i in list(zip(df['City'], df['Country'])):
    if type(i[0]) == float:
        try:
            loc = geolocator.geocode(i[1])
            geo.append((loc.latitude, loc.longitude))
        except:
            geo.append(np.nan)
    else:
        try:
            location = ', '.join(i)
            loc = geolocator.geocode(location)
            geo.append((loc.latitude, loc.longitude))
        except:
            geo.append(np.nan)
            
df['Location'] = geo

def true_false_converter(x):
    if x == True:
        return 'Yes'
    elif x == 'yes':
        return 'Yes'
    elif x == False:
        return 'No'
    elif x == 'no':
        return 'No'
    elif x == 'dn':
        return np.nan
    else:
        return x

df[df.columns[0:16]] = df[df.columns[0:16]].applymap(true_false_converter)

capitalizer = lambda x: x.title()

df['Gender'] = df['Gender'].map(capitalizer)

def education_converter(x):
    if x == 1:
        return 'Some Primary School'
    elif x == 2:
        return 'Finished Primary School'
    elif x == 3:
        return 'Finished Secondary School'
    elif x == 4:
        return 'Beyond Secondary School'
    
df['Education'] = df['Education'].map(education_converter)

df.drop(['City', 'Country', 'Location'], axis=1).to_csv('clean_data/survey_data.csv', index=True)

df[df.columns[22:]].to_csv('clean_data/survey_location_data.csv', index=True)

len(df.columns.values) * 5623

## Clean UReport Data

ur_df = pd.read_csv('raw_data/UN Women poll results - Hoja 1.csv', header=1)

ur_df = ur_df.rename(index=str, columns={"Unnamed: 0": "Country", 
                                         "Unnamed: 1": "Date",
                                         "Unnamed: 2": "Total Polled",
                                         "Unnamed: 3": "Responded",
                                         "%YES": "Believes In Pay Equity",
                                         "%NO": "Does Not Belive In Pay Equity",
                                         "Fear of violence %": "Fear Of Violence Is A Barrier",
                                         "Legal obstacles %": "Laws Against Women Working Is A Barrier",
                                         "% Family responsibilities ": "Family Responsibilities Are A Barrier",
                                         "Lack of skills %": "Lack Of Experience Or Skills Is A Barrier",
                                         "Traditional views of women": "Traditional Views Of Women Are A Barrier",
                                         "Lack of digital connectivity": "Lack Of ICT Access Is A Barrier",
                                         "Your own": "Other Things Are Barriers",
                                         "Good education": "Education Helps Women",
                                         "Women rolemodels": "Role Models Help Women",
                                         "Family planning & contraception": "Family Planning Helps Women",
                                         "Maternity pay & childcare": "Maternity Pay And Childcare Helps Women",
                                         "access to banking": "Access To Banking Helps Women",
                                         "women's groups": "Womens Groups And Charities Helps Women",
                                         "your own": "Other Things Help Women",
                                         }
                    )

def string_to_float(x):
    if type(x) != float:
        return x.replace('.', '').replace(',', '.')
    else:
        return x

ur_df['Total Polled'] = [string_to_float(x) for x in ur_df['Total Polled']]
ur_df['Responded'] = [string_to_float(x) for x in ur_df['Responded']]

ur_df = ur_df[ur_df.Country != 'Global']
ur_df = ur_df[ur_df.Country != 'Central African Republic']
ur_df

## Add Extra Columns

Income = {'Swaziland':'Middle',
          'Guinea':'Low',
          'Brazil':'Middle',
          'Burkina Faso':'Low',
          'Cameroon':'Middle',
          'Ireland':'Very High',
          'Mexico':'High',
          'Ukraine':'Middle',
          'Zimbabwe':'Low',
          'Liberia':'Low',
          'Pakistan':'Middle',
          'Malaysia':'Middle',
          'Indonesia':'Middle',
          'Nigiera':'Middle',
         }

Region = {'Swaziland':'Sub-Saharan Africa',
          'Guinea':'Sub-Saharan Africa',
          'Brazil':'Latin America and the Caribbean',
          'Burkina Faso':'Sub-Saharan Africa',
          'Cameroon':'Sub-Saharan Africa',
          'Ireland':'Europe',
          'Mexico':'Central America',
          'Ukraine':'Eastern Europe and central Asia',
          'Zimbabwe':'Sub-Saharan Africa',
          'Liberia':'Sub-Saharan Africa',
          'Pakistan':'Asia',
          'Malaysia':'Asia',
          'Indonesia':'South East Asia',
          'Nigiera':'Sub-Saharan Africa',
         }

df_income = pd.DataFrame()

df_income = df_income.from_dict(Income, orient='index')
df_income['Income Level'] = df_income[0]
df_income = df_income.drop(df_income.columns[[0]], axis=1)
ur_df = ur_df.merge(df_income, left_on='Country', right_index=True)

df_region = df_income.from_dict(Region, orient='index')
df_region['Region'] = df_region[0]
df_region = df_region.drop(df_region.columns[[0]], axis=1)
ur_df = ur_df.merge(df_region, left_on='Country', right_index=True)
ur_df

ur_df.to_csv('clean_data/ureport_country_data.csv')

## Create Aggregate Ranking Barriers

barriers = ["Fear Of Violence Is A Barrier",
            "Laws Against Women Working Is A Barrier",
            "Family Responsibilities Are A Barrier",
            "Lack Of Experience Or Skills Is A Barrier",
            "Traditional Views Of Women Are A Barrier",
            "Lack Of ICT Access Is A Barrier",
            "Other Things Are Barriers"]

meaner = lambda x: x.str.replace('%','').astype(int).mean()

ur_df_barriers_ranked = ur_df[barriers].apply(meaner).to_frame(name='Average Percent Agreed')

ur_df_barriers_ranked['Barrier'] = ur_df_barriers_ranked.index
ur_df_barriers_ranked

ur_df_barriers_ranked.sort_values('Average Percent Agreed', ascending=False).to_csv('clean_data/barriers_ranked.csv')

## Create Aggregate Ranking Barriers

actions = ["Education Helps Women",
            "Role Models Help Women",
            "Family Planning Helps Women",
            "Maternity Pay And Childcare Helps Women",
            "Access To Banking Helps Women",
            "Womens Groups And Charities Helps Women",
            "Other Things Help Women"]

ur_df[actions]

meaner = lambda x: x.str.replace('%','').astype(int).mean()

ur_df_actions_ranked = ur_df[actions].dropna(how='all').apply(meaner).to_frame(name='Average Percent Agreed')

ur_df_actions_ranked['Action'] = ur_df_actions_ranked.index

ur_df_actions_ranked['Action'] = ur_df_actions_ranked['Action'].str.replace('Helps Women', '')
ur_df_actions_ranked['Action'] = ur_df_actions_ranked['Action'].str.replace('Help Women', '')

ur_df_actions_ranked

ur_df_actions_ranked.sort_values('Average Percent Agreed', ascending=False).to_csv('clean_data/actions_ranked.csv')
