==> data/step_3_aggregates.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('clean_data/ureport_country_data.csv')
df = df.drop('Unnamed: 0', axis=1)

barriers = ["Fear Of Violence Is A Barrier",
            "Laws Against Women Working Is A Barrier",
            "Family Responsibilities Are A Barrier",
            "Lack Of Experience Or Skills Is A Barrier",
            "Traditional Views Of Women Are A Barrier",
            "Lack Of ICT Access Is A Barrier",
            "Other Things Are Barriers"]

actions = ['Education ', 
           'Role Models ', 
           'Family Planning ',
           'Maternity Pay And Childcare ', 
           'Access To Banking ',
           'Womens Groups And Charities ', 
           'Other Things ']

x = np.nan

def cleaner(x):
    if type(x) == str:
        try:
            return int(x.replace('%', ''))
        except:
            return x
    elif type(x) == float:
        return x
    else:
        return x

df_clean = df.applymap(cleaner)

df_region = df_clean.groupby('Region').mean()
df_region.columns = [x.replace('Help Women', '').replace('Helps Women', '') for x in df_region.columns.values]
df_region['Region'] = df_region.index
df_region = df_region.round(2)

df_income = df_clean.groupby('Income Level').mean()
df_income.columns = [x.replace('Help Women', '').replace('Helps Women', '') for x in df_income.columns.values]
df_income['Income Level'] = df_income.index
df_income = df_income.round(2)

df_region_barriers = df_region[barriers].transpose()
df_region_barriers['Barriers'] = df_region_barriers.index
df_region_barriers.to_csv('clean_data/ureport_region_barriers.csv', index=False)

df_region_actions = df_region[actions].transpose()
df_region_actions['Actions'] = df_region_actions.index
df_region_actions.to_csv('clean_data/ureport_region_actions.csv', index=False)

df_income_barriers = df_income[barriers].transpose()
df_income_barriers['Barriers'] = df_income_barriers.index
df_income_barriers.to_csv('clean_data/ureport_income_barriers.csv', index=False)

df_income_actions = df_income[actions].transpose()
df_income_actions['Actions'] = df_income_actions.index
df_income_actions.to_csv('clean_data/ureport_income_actions.csv', index=False)

df_income_actions

df_income_actions[['Low', 'Actions']].to_csv('clean_data/ureport_income_actions_low.csv', index=False)

df_region_actions

df_region_actions[['Latin America and the Caribbean', 'Actions']].to_csv('clean_data/ureport_region_actions_la.csv', index=False)

df_region_barriers[['Europe', 'Barriers']].to_csv('clean_data/ureport_region_barriers_europe.csv', index=False)

df_region_actions
